refactor: extract WysiwygEditor component and move MD↔HTML converters to utils/markdown.js
