+++ src/components/EditModal/index.jsx
@@ -1,369 +1,12 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React, { useState, useEffect } from 'react';
 import ReactDOM from 'react-dom';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import { fetchRawFile, createEditPR } from '@site/src/utils/github';
-import { splitFrontmatter, inlineMd } from '@site/src/utils/markdown';
+import { splitFrontmatter, mdToHtml, htmlToMd } from '@site/src/utils/markdown';
+import { WysiwygEditor } from '@site/src/components/WysiwygEditor';
 import styles from './index.module.css';
 
-export function mdToHtml(md) {
-  const lines = md.split('\n');
-  const out = [];
-  let i = 0;
-  while (i < lines.length) {
-    const line = lines[i];
-
-    if (line.startsWith('```')) {
-      const codeLines = [];
-      i++;
-      while (i < lines.length && !lines[i].startsWith('```')) {
-        codeLines.push(lines[i++]);
-      }
-      out.push(`<pre><code>${codeLines.join('\n')}</code></pre>`);
-      i++;
-      continue;
-    }
-
-    const heading = line.match(/^(#{1,6})\s+(.+)$/);
-    if (heading) {
-      const lvl = heading[1].length;
-      out.push(`<h${lvl}>${inlineMd(heading[2])}</h${lvl}>`);
-      i++;
-      continue;
-    }
-
-    if (line.match(/^[-*+]\s/)) {
-      const items = [];
-      while (i < lines.length && lines[i].match(/^[-*+]\s/)) {
-        items.push(`<li>${inlineMd(lines[i].replace(/^[-*+]\s/, ''))}</li>`);
-        i++;
-      }
-      out.push(`<ul>${items.join('')}</ul>`);
-      continue;
-    }
-
-    if (line.match(/^\d+\.\s/)) {
-      const items = [];
-      while (i < lines.length && lines[i].match(/^\d+\.\s/)) {
-        items.push(`<li>${inlineMd(lines[i].replace(/^\d+\.\s/, ''))}</li>`);
-        i++;
-      }
-      out.push(`<ol>${items.join('')}</ol>`);
-      continue;
-    }
-
-    if (line.match(/^>+\s/)) {
-      const bqLines = [];
-      while (i < lines.length && lines[i].match(/^>+\s?/)) {
-        bqLines.push(inlineMd(lines[i].replace(/^>+\s?/, '')));
-        i++;
-      }
-      out.push(`<blockquote><p>${bqLines.join('<br>')}</p></blockquote>`);
-      continue;
-    }
-
-    if (line.trim() === '') {
-      i++;
-      continue;
-    }
-
-    const paraLines = [];
-    while (
-      i < lines.length &&
-      lines[i].trim() !== '' &&
-      !lines[i].match(/^#{1,6}\s/) &&
-      !lines[i].match(/^[-*+]\s/) &&
-      !lines[i].match(/^\d+\.\s/) &&
-      !lines[i].match(/^>/) &&
-      !lines[i].startsWith('```')
-    ) {
-      paraLines.push(inlineMd(lines[i]));
-      i++;
-    }
-    if (paraLines.length > 0) {
-      out.push(`<p>${paraLines.join('<br>')}</p>`);
-    }
-  }
-  return out.join('\n');
-}
-
-function nodeToMd(node) {
-  let out = '';
-  for (const child of node.childNodes) {
-    if (child.nodeType === 3) {
-      out += child.textContent;
-      continue;
-    }
-    if (child.nodeType !== 1) continue;
-    const tag = child.tagName.toLowerCase();
-    const inner = nodeToMd(child);
-    switch (tag) {
-      case 'h1': out += `\n\n# ${inner.trim()}\n\n`; break;
-      case 'h2': out += `\n\n## ${inner.trim()}\n\n`; break;
-      case 'h3': out += `\n\n### ${inner.trim()}\n\n`; break;
-      case 'h4': out += `\n\n#### ${inner.trim()}\n\n`; break;
-      case 'h5': out += `\n\n##### ${inner.trim()}\n\n`; break;
-      case 'h6': out += `\n\n###### ${inner.trim()}\n\n`; break;
-      case 'strong': case 'b': out += `**${inner}**`; break;
-      case 'em': case 'i': out += `*${inner}*`; break;
-      case 'u': out += `__${inner}__`; break;
-      case 's': case 'del': case 'strike': out += `~~${inner}~~`; break;
-      case 'code':
-        if (child.parentElement?.tagName.toLowerCase() === 'pre') out += inner;
-        else out += `\`${inner}\``;
-        break;
-      case 'pre': out += `\n\n\`\`\`\n${inner}\n\`\`\`\n\n`; break;
-      case 'a': out += `[${inner}](${child.getAttribute('href') || ''})`; break;
-      case 'img': out += `![${child.getAttribute('alt') || ''}](${child.getAttribute('src') || ''})`; break;
-      case 'iframe': {
-        const src = child.getAttribute('src') || '';
-        const w = child.getAttribute('width') || '560';
-        const h = child.getAttribute('height') || '315';
-        out += `\n\n<iframe width="${w}" height="${h}" src="${src}" frameborder="0" allowfullscreen style="max-width:100%"></iframe>\n\n`;
-        break;
-      }
-      case 'video': {
-        const src = child.getAttribute('src') || '';
-        out += `\n\n<video src="${src}" controls style="max-width:100%;height:auto"></video>\n\n`;
-        break;
-      }
-      case 'p': out += `\n\n${inner}\n\n`; break;
-      case 'br': out += '\n'; break;
-      case 'blockquote': out += `\n\n> ${inner.trim().replace(/\n/g, '\n> ')}\n\n`; break;
-      case 'ul': {
-        for (const li of child.querySelectorAll(':scope > li')) {
-          out += `\n- ${nodeToMd(li).trim()}`;
-        }
-        out += '\n';
-        break;
-      }
-      case 'ol': {
-        let idx = 1;
-        for (const li of child.querySelectorAll(':scope > li')) {
-          out += `\n${idx}. ${nodeToMd(li).trim()}`;
-          idx++;
-        }
-        out += '\n';
-        break;
-      }
-      case 'li': out += inner; break;
-      case 'span': {
-        const colorMatch = (child.getAttribute('style') || '').match(/color\s*:\s*([^;]+)/i);
-        out += colorMatch
-          ? `<span style="color:${colorMatch[1].trim()}">${inner}</span>`
-          : inner;
-        break;
-      }
-      case 'div': out += `\n${inner}\n`; break;
-      default: out += inner;
-    }
-  }
-  return out;
-}
-
-export function htmlToMd(html) {
-  const div = document.createElement('div');
-  div.innerHTML = html;
-  return nodeToMd(div).replace(/\n{3,}/g, '\n\n').trim();
-}
-
-/* ── WYSIWYG Editor ─────────────────────────────────────────────────── */
-
-function ToolBtn({ title, onAction, children }) {
-  return (
-    <button
-      type="button"
-      title={title}
-      className={styles.toolBtn}
-      onMouseDown={(e) => { e.preventDefault(); onAction(); }}
-    >
-      {children}
-    </button>
-  );
-}
-
-export function WysiwygEditor({ initialHtml, onChange }) {
-  const editorRef = useRef(null);
-  const colorRef = useRef(null);
-  const imgInputRef = useRef(null);
-  const vidInputRef = useRef(null);
-  const initialized = useRef(false);
-
-  useEffect(() => {
-    if (editorRef.current && !initialized.current) {
-      editorRef.current.innerHTML = initialHtml || '';
-      initialized.current = true;
-      document.execCommand('defaultParagraphSeparator', false, 'p');
-    }
-  }, [initialHtml]);
-
-  const exec = useCallback((cmd, val = null) => {
-    document.execCommand(cmd, false, val);
-    editorRef.current?.focus();
-  }, []);
-
-  const handleInput = useCallback(() => {
-    onChange(editorRef.current?.innerHTML || '');
-  }, [onChange]);
-
-  const insertLink = useCallback(() => {
-    const url = window.prompt('Enter link URL:');
-    if (url) exec('createLink', url);
-  }, [exec]);
-
-  const handleColor = useCallback((e) => {
-    exec('foreColor', e.target.value);
-  }, [exec]);
-
-  const handleImageFile = useCallback((e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    const defaultAlt = file.name.replace(/\.[^.]+$/, '');
-    const alt = window.prompt('Alt text for the image:', defaultAlt) ?? defaultAlt;
-    const reader = new FileReader();
-    reader.onload = (ev) => {
-      editorRef.current?.focus();
-      document.execCommand('insertHTML', false,
-        `<img src="${ev.target.result}" alt="${alt}" style="max-width:100%;height:auto">`
-      );
-      onChange(editorRef.current?.innerHTML || '');
-    };
-    reader.readAsDataURL(file);
-    e.target.value = '';
-  }, [onChange]);
-
-  const handleVideoInsert = useCallback(() => {
-    const url = window.prompt('Enter video URL (YouTube, Vimeo, or direct .mp4/.webm):');
-    if (!url) return;
-    let html;
-    const ytMatch = url.match(/(?:youtube\.com\/(?:watch\?v=|embed\/)|youtu\.be\/)([a-zA-Z0-9_-]{11})/);
-    const vimeoMatch = url.match(/vimeo\.com\/(?:video\/)?(\d+)/);
-    if (ytMatch) {
-      html = `<iframe width="560" height="315" src="https://www.youtube.com/embed/${ytMatch[1]}" frameborder="0" allowfullscreen style="max-width:100%"></iframe>`;
-    } else if (vimeoMatch) {
-      html = `<iframe width="560" height="315" src="https://player.vimeo.com/video/${vimeoMatch[1]}" frameborder="0" allowfullscreen style="max-width:100%"></iframe>`;
-    } else {
-      html = `<video src="${url}" controls style="max-width:100%;height:auto"></video>`;
-    }
-    editorRef.current?.focus();
-    document.execCommand('insertHTML', false, html);
-    onChange(editorRef.current?.innerHTML || '');
-  }, [onChange]);
-
-  const handleVideoFile = useCallback((e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    if (file.size > 10 * 1024 * 1024) {
-      window.alert('Video file exceeds 10 MB. Please use the 🎬 URL button to link to a hosted video instead.');
-      e.target.value = '';
-      return;
-    }
-    const reader = new FileReader();
-    reader.onload = (ev) => {
-      editorRef.current?.focus();
-      document.execCommand('insertHTML', false,
-        `<video src="${ev.target.result}" controls style="max-width:100%;height:auto"></video>`
-      );
-      onChange(editorRef.current?.innerHTML || '');
-    };
-    reader.readAsDataURL(file);
-    e.target.value = '';
-  }, [onChange]);
-
-  return (
-    <div className={styles.editorWrap}>
-      <div className={styles.toolbar}>
-        <ToolBtn title="Bold (Ctrl+B)" onAction={() => exec('bold')}>
-          <strong>B</strong>
-        </ToolBtn>
-        <ToolBtn title="Italic (Ctrl+I)" onAction={() => exec('italic')}>
-          <em>I</em>
-        </ToolBtn>
-        <ToolBtn title="Underline (Ctrl+U)" onAction={() => exec('underline')}>
-          <u>U</u>
-        </ToolBtn>
-        <ToolBtn title="Strikethrough" onAction={() => exec('strikeThrough')}>
-          <s>S</s>
-        </ToolBtn>
-        <span className={styles.toolSep} />
-        <ToolBtn title="Heading 1" onAction={() => exec('formatBlock', 'h1')}>H1</ToolBtn>
-        <ToolBtn title="Heading 2" onAction={() => exec('formatBlock', 'h2')}>H2</ToolBtn>
-        <ToolBtn title="Heading 3" onAction={() => exec('formatBlock', 'h3')}>H3</ToolBtn>
-        <span className={styles.toolSep} />
-        <ToolBtn title="Bullet list" onAction={() => exec('insertUnorderedList')}>
-          ≡•
-        </ToolBtn>
-        <ToolBtn title="Numbered list" onAction={() => exec('insertOrderedList')}>
-          ≡1
-        </ToolBtn>
-        <ToolBtn title="Blockquote" onAction={() => exec('formatBlock', 'blockquote')}>
-          "
-        </ToolBtn>
-        <span className={styles.toolSep} />
-        <ToolBtn title="Insert link" onAction={insertLink}>🔗</ToolBtn>
-        <ToolBtn title="Remove link" onAction={() => exec('unlink')}>🔗̶</ToolBtn>
-        <span className={styles.toolSep} />
-        {/* Image insert — label wraps hidden file input so click is a trusted user gesture */}
-        <label
-          className={styles.toolBtn}
-          title="Insert image from file"
-          onMouseDown={(e) => e.preventDefault()}
-        >
-          🖼
-          <input
-            ref={imgInputRef}
-            type="file"
-            accept="image/*"
-            style={{ display: 'none' }}
-            onChange={handleImageFile}
-          />
-        </label>
-        <ToolBtn title="Insert video (YouTube, Vimeo, or direct URL)" onAction={handleVideoInsert}>🎬</ToolBtn>
-        <label
-          className={styles.toolBtn}
-          title="Insert video from file (max 10 MB)"
-          onMouseDown={(e) => e.preventDefault()}
-        >
-          📹
-          <input
-            ref={vidInputRef}
-            type="file"
-            accept="video/*"
-            style={{ display: 'none' }}
-            onChange={handleVideoFile}
-          />
-        </label>
-        <span className={styles.toolSep} />
-        <label className={styles.colorLabel} title="Text color">
-          <span className={styles.colorIcon}>A</span>
-          <input
-            ref={colorRef}
-            type="color"
-            className={styles.colorInput}
-            defaultValue="#e74c3c"
-            onChange={handleColor}
-            onMouseDown={(e) => e.stopPropagation()}
-          />
-        </label>
-        <ToolBtn title="Remove formatting" onAction={() => exec('removeFormat')}>
-          Tx
-        </ToolBtn>
-      </div>
-      <div
-        ref={editorRef}
-        contentEditable
-        suppressContentEditableWarning
-        className={styles.editorContent}
-        onInput={handleInput}
-        onKeyDown={(e) => {
-          if (e.key === 'Tab') {
-            e.preventDefault();
-            exec('insertHTML', '&nbsp;&nbsp;');
-          }
-        }}
-      />
-    </div>
-  );
-}
+export { mdToHtml, htmlToMd, WysiwygEditor };
 
 /* ── Form field helper ───────────────────────────────────────────────── */
